backend/requirements.txt
docker==7.1.0
fastapi==0.115.5
joblib==1.4.2
numpy==2.1.3
pandas==2.2.3
passlib==1.7.4
prefect==3.1.4
pydantic==2.10.0
pydantic_core==2.27.0
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pyjwt==2.10.0
scikit-learn==1.5.2
seaborn==0.13.2
SQLAlchemy==2.0.36
starlette==0.41.3
urllib3==2.2.3
uvicorn==0.32.1